back/requirements.txt
Flask==3.0.3
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
Flask-Cors==5.0